fix(review): address pre-ship code-review findings
(1) studio scrapbook-file route — validate `entryId` query param against
    UUID regex before passing to `readSidecar`. Without this guard, an
    `entryId` of `../../../foo` lets `node:path` join collapse `..`
    segments and probe arbitrary on-disk locations. No data is leaked
    (ENOENT becomes 404) but the containment posture matches the rest
    of the scrapbook helpers now.

(2) annotations.ts was 549 lines (49 over the 300–500 cap). Extracted
    the resolved-footer rendering into `resolved-footer.ts` and the
    pure folds (`resolvedCommentIds`, `latestAddressByCommentId`) into
    `annotation-folding.ts`. The controller is now 490 lines.

Stale-geometry concern flagged in review (sidebar.getBoundingClientRect
called synchronously after unstowMarginalia) was empirically refuted —
Playwright verification with the drawer initially stowed showed the
composer landing at zero delta from the selection. `getBoundingClientRect`
is a forced synchronous layout call, so the rect reflects post-style-
change geometry. No code change needed.
